chore(android): fix test

+++ measure-android/measure/src/test/java/sh/measure/android/storage/DatabaseTest.kt
@@ -134,7 +134,7 @@ class DatabaseTest {
         )
         assertEquals(
             expectedEvent.attachmentsSize,
-            cursor.getInt(cursor.getColumnIndex(EventTable.COL_ATTACHMENT_SIZE)),
+            cursor.getLong(cursor.getColumnIndex(EventTable.COL_ATTACHMENT_SIZE)),
         )
     }
